Pivotal ID # 187384535: Backend Kotlin upgrade (#824)
https://www.pivotaltracker.com/story/show/187384535
- added deteck
- added klint
- reformat code
- using spring resolution

+++ submission/persistence-filesystem/src/main/kotlin/ac/uk/ebi/biostd/persistence/filesystem/service/StorageService.kt
@@ -22,7 +22,10 @@ class StorageService(
     private val nfsFilesService: NfsFilesService,
     private val serializationService: ExtSerializationService,
 ) : FileStorageService {
-    override suspend fun persistSubmissionFile(sub: ExtSubmission, file: ExtFile): ExtFile =
+    override suspend fun persistSubmissionFile(
+        sub: ExtSubmission,
+        file: ExtFile,
+    ): ExtFile =
         when (sub.storageMode) {
             FIRE -> fireFilesService.persistSubmissionFile(sub, file)
             NFS -> nfsFilesService.persistSubmissionFile(sub, file)
@@ -40,7 +43,10 @@ class StorageService(
         }
     }
 
-    override suspend fun deleteSubmissionFile(sub: ExtSubmission, file: ExtFile) = when (sub.storageMode) {
+    override suspend fun deleteSubmissionFile(
+        sub: ExtSubmission,
+        file: ExtFile,
+    ) = when (sub.storageMode) {
         FIRE -> {
             fireFilesService.deleteSubmissionFile(sub, file)
             fireFilesService.deleteFtpFile(sub, file)
@@ -60,8 +66,9 @@ class StorageService(
         deleteEmptyFolders(sub)
     }
 
-    override suspend fun deleteEmptyFolders(sub: ExtSubmission) = when (sub.storageMode) {
-        FIRE -> fireFilesService.deleteEmptyFolders(sub)
-        NFS -> nfsFilesService.deleteEmptyFolders(sub)
-    }
+    override suspend fun deleteEmptyFolders(sub: ExtSubmission) =
+        when (sub.storageMode) {
+            FIRE -> fireFilesService.deleteEmptyFolders(sub)
+            NFS -> nfsFilesService.deleteEmptyFolders(sub)
+        }
 }
